+++ tests/testthat/test-16-prefab_models.R
@@ -11,7 +11,7 @@ main <- "./testing"
 
 
   fill_data(main = main)
-  keepers <- c("moon", "DM", "ntraps")
+  keepers <- c("newmoonnumber", "DM", "ntraps")
 
 
   all <- read_rodents_table(main = main, "all")
@@ -179,7 +179,7 @@ test_that(desc = "jags_RW model", {
 
   rjc <- runjags_control(adapt = 100, burnin = 100, sample = 100)
   expect_message(f_a <- jags_RW(main = main, control_runjags = rjc,
-                                dataset = "dm_controls", quiet = FALSE))
+                                dataset = "Controls", quiet = FALSE))
   expect_is(f_a, "list")
 
 })
@@ -196,7 +196,7 @@ test_that(desc = "jags_logistic model", {
 
   rjc <- runjags_control(adapt = 100, burnin = 100, sample = 100)
   expect_message(f_a <- jags_logistic(main = main, control_runjags = rjc,
-                                dataset = "dm_controls", quiet = FALSE))
+                                dataset = "Controls", quiet = FALSE))
   expect_is(f_a, "list")
 
 
@@ -214,7 +214,45 @@ test_that(desc = "jags_logistic_covariates model", {
 
   rjc <- runjags_control(adapt = 100, burnin = 100, sample = 100)
   expect_message(f_a <- jags_logistic_covariates(main = main, control_runjags = rjc,
-                                dataset = "dm_controls", quiet = FALSE))
+                                dataset = "Controls", quiet = FALSE))
+  expect_is(f_a, "list")
+
+
+})
+
+
+test_that(desc = "jags_logistic_competition model", {
+
+  # download is held back on cran
+
+    skip_on_cran() 
+
+  # check for JAGS installation on system
+
+    check_jags() 
+
+  rjc <- runjags_control(adapt = 100, burnin = 100, sample = 100)
+  expect_message(f_a <- jags_logistic_competition(main = main, control_runjags = rjc,
+                                dataset = "Controls", quiet = FALSE))
+  expect_is(f_a, "list")
+
+
+})
+
+
+test_that(desc = "jags_logistic_competition_covariates model", {
+
+  # download is held back on cran
+
+    skip_on_cran() 
+
+  # check for JAGS installation on system
+
+    check_jags() 
+
+  rjc <- runjags_control(adapt = 100, burnin = 100, sample = 100)
+  expect_message(f_a <- jags_logistic_competition_covariates(main = main, control_runjags = rjc,
+                                dataset = "Controls", quiet = FALSE))
   expect_is(f_a, "list")