+++ NEWS.md
@@ -2,6 +2,19 @@
 
 Version numbers follow [Semantic Versioning](https://semver.org/).
 
+# portalcasting 0.45.0
+*In Progress*
+
+### Removing `dm_controls` as a separate dataset
+* Following the removal of the `interp` datasets
+* Now model controls indicate species to which their applied
+
+### New models
+* `jags_logistic_competition`
+* `jags_logistic_covariates`
+* `jags_logistic_competition_covariates`
+
+
 # [portalcasting 0.44.0](https://github.com/weecology/portalcasting/releases/tag/v0.44.0)
 *2022-10-27*